Added location details to booking history.

+++ frontend/src/app/ongoing-bookings/ongoing-bookings.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { Booking } from '../models/booking';
 import { BookingService } from '../service/booking.service';
 import { UserService } from '../service/user.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-ongoing-bookings',
@@ -12,7 +14,10 @@ export class OngoingBookingsComponent {
   public bookings: Booking[] = [];
   private userId: number = 0;
 
-  constructor(private bookingService: BookingService) {
+  constructor(
+    private bookingService: BookingService,
+    private dialog: MatDialog
+  ) {
     this.userId = parseInt(localStorage.getItem('currentUserId')!);
     this.bookingService.getBookingsForUser(this.userId).subscribe({
       next: (response) => {
@@ -25,15 +30,24 @@ export class OngoingBookingsComponent {
   }
 
   deleteBooking(bookingID: number) {
-    this.bookingService.deleteBooking(bookingID).subscribe({
-      next: (response) => {
-        this.bookings = this.bookings.filter(
-          (booking) => booking.id !== bookingID
-        );
-      },
-      error: (error) => {
-        console.log(error);
-      },
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '350px',
+      data: { message: 'Delete this entry?' },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.bookingService.deleteBooking(bookingID).subscribe({
+          next: (response) => {
+            this.bookings = this.bookings.filter(
+              (booking) => booking.id !== bookingID
+            );
+          },
+          error: (error) => {
+            console.log(error);
+          },
+        });
+      }
     });
   }
